+++ src/lib/entities/auth/ConfirmEmailDto.ts
@@ -1,0 +1,4 @@
+export type ConfirmEmailDto = {
+  tokenHash: string;
+  type: string;
+};