+++ src/lib/entities/auth/ForgotPasswordDto.ts
@@ -1,0 +1,4 @@
+export type ForgotPasswordDto<TExtra extends object = object> = {
+  email: string;
+  redirectTo?: string;
+} & TExtra;