+++ src/lib/entities/auth/ResetPasswordDto.ts
@@ -1,0 +1,15 @@
+export type ResetPasswordByAccessTokenDto = {
+  accessToken: string;
+  refreshToken?: string;
+  newPassword: string;
+};
+
+export type ResetPasswordByTokenHashDto = {
+  tokenHash: string;
+  type: string;
+  newPassword: string;
+};
+
+export type ResetPasswordDto =
+  | ResetPasswordByAccessTokenDto
+  | ResetPasswordByTokenHashDto;